add json payload middleware and json error middleware
